Modified GraphView
Set the GraphView to full screen instead of a smaller window. Also, had a wierd issue where Android Studio on my device was reading the activity_view files incorrectly and was displaying code from another part of the project instead of the xml code. It must be a local issue because GitHub still has the correct xml code. Weird.